pyproject.toml
[tool.poetry]
name = "ml-project"
version = "0.1.0"
description = "Project for the AI Center Projects in Machine Learning Research course"
authors = [
  "András Geiszl <geiszla@gmail.com>",
  "Gian Hess <gihess@student.ethz.ch>",
  "Yunxiang Guo <yunguo@student.ethz.ch>",
  "Yves Bicker <bickery@student.ethz.ch>",
]
license = "UNLICENSED"
readme = "README.md"
repository = "https://github.com/xiaoxiaoshikui/Machine-Learning-Project-for-ETH-AI-Center"
packages = [{ include = "ml_project" }]

[tool.poetry.scripts]
gridworld = "ml_project.gridWorld:main"
generate = "ml_project.rl.generate_videos_and_transitions:main"
train = "ml_project.rl.train_rl_agent:main"

[tool.poetry.dependencies]
python = "~3.10"
flask = "^2.3.2"
numpy = "^1.24.3"
gymnasium = "^0.28.1"
# Required to build mujoco-py
cython = "^0.29.34"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
conda-lock = "^1.4.0"
pre-commit = "^3.3.1"
pylint = "^2.17.3"
pyright = "^1.1.306"
ruff = "^0.0.264"
# Need to fix this, otherwise Poetry upgrades it and breaks itself
urllib3 = "<2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.pyright]
pythonVersion = "3.10"
pythonPlatform = "All"

include = ["ml_project"]
exclude = ["**/__pycache__"]

typeCheckingMode = "basic"

useLibraryCodeForTypes = true
stubPath = "python-type-stubs"
